+++ CMakeLists.txt
@@ -1,15 +1,16 @@
 cmake_minimum_required(VERSION 3.26)
-project(untitled4)
+project(loq_v1)
 
 set(CMAKE_CXX_STANDARD 17)
 
 
-add_library(untitled4 SHARED
-        src/block/iblock.cpp
-        src/block/block.h src/library.h
-        src/block/lblock.cpp
-        src/validate/endable.cpp
+add_library(loq_v1 SHARED
+        src/block/block.cpp
+        src/block/block.h
+	src/library.cpp
+	src/library.h
         src/point.cpp
         src/state.cpp
         src/actions/move.cpp
-        src/actions/place.cpp)
+        src/actions/place.cpp
+	src/validate/endable.cpp)
