+++ src/library.cpp
@@ -1,0 +1,5 @@
+unsigned char abs(unsigned char x) { return x > 0 ? x : -x; }
+
+unsigned char max(unsigned char x, unsigned char y) { return x > y ? x : y; }
+
+unsigned char min(unsigned char x, unsigned char y) { return x < y ? x : y; }
